requirements_extra.txt
apache-airflow-providers-alibaba
apache-airflow-providers-elasticsearch
apache-airflow-providers-mysql[amazon]
apache-airflow-providers-ssh
openpyxl
lxml
html5lib
beautifulsoup4
stringcase
elasticsearch==7.10.0
pangres